backend/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
termcolor>=2.3.0 